PockeTutor/Backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
itsdangerous==2.2.0
python-docx==1.1.2
reportlab==4.2.2
openai
requests
pdfminer.six
python-docx
reportlab
Flask-JWT-Extended
